JWT sent to angular
