+++ angular/SIG/src/app/Components/accountdetail/accountdetail.component.ts
