+++ angular/SIG/src/app/Components/orders/orders.component.html
@@ -11,46 +11,44 @@
 </div>
 
 <div class="card text-center">
-    <h5 class="card-header">Brokerage ...7844</h5>
-    <div class="card-body">
-      <h6 class="card-title">Available Cash: $50,000,000.00</h6>
-    </div>
+  <h5 class="card-header">Brokerage ...7844</h5>
+  <div class="card-body">
+    <h6 class="card-title">Available Cash: $50,000,000.00</h6>
+  </div>
 </div>
 
 
-<form style="text-align: center;">    
-    <div class="form-row">      
-      <div class="form-group col-md-4">
-        <label for="inputTransactionType">Transaction Type</label>
-        <select id="inputTransactionType" class="form-control">
-          <option selected>Stocks/ETFs</option>
-          <option>Mutual Funds</option>
-          <option>Bonds</option>
-        </select>
-      </div>
-      <div class="form-group col-md-4">
-          <label for="inputSymbol">Symbol</label>
-          <input type="text" class="form-control" id="inputSymbol">
-      </div>      
-    </div>    
-    <div class="card text-center">
-        <div class="card-header">
-          GOOG
-        </div>
-        <div class="card-body">
-          <h5 class="card-title">1,103</h5>
-        </div>
-        <div class="card-footer text-muted">
-          Volume 2.00M
-        </div>
+<form style="text-align: center;">
+  <div class="form-row">
+    <div class="form-group col-md-4">
+      <label for="inputTransactionType">Transaction Type</label>
+      <select id="inputTransactionType" class="form-control">
+        <option selected>Stocks/ETFs</option>
+        <option>Mutual Funds</option>
+        <option>Bonds</option>
+      </select>
     </div>
-    <button type="submit" class="btn btn-primary">Place Order</button>
-    <a href="#" class="btn btn-primary">Place Order</a>
+    <div class="form-group col-md-4">
+      <label for="inputSymbol">Symbol</label>
+      <input type="text" class="form-control" id="inputSymbol">
+    </div>
+  </div>
+  <div class="card text-center">
+    <div class="card-header">
+      GOOG
+    </div>
+    <div class="card-body">
+      <h5 class="card-title">1,103</h5>
+    </div>
+    <div class="card-footer text-muted">
+      Volume 2.00M
+    </div>
+  </div>
+  <button type="submit" class="btn btn-primary">Place Order</button>
+  <a href="#" class="btn btn-primary">Place Order</a>
 </form>
 
 
 <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
-
-
