+++ angular/SIG/src/app/Services/balance.service.ts
@@ -2,19 +2,19 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http';
 
 @Injectable({
-  providedIn: 'root'
+	providedIn: 'root'
 })
 export class BalanceService {
-  // I need this service to return balance(s) from our DB.
-  constructor(private http: HttpClient) { }
+	// I need this service to return balance(s) from our DB.
+	constructor(private http: HttpClient) { }
 
-  httpOptions = {
-    headers: new HttpHeaders({ 'Content-Type': 'application/json' })
-  };
+	httpOptions = {
+		headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+	};
 
-  // this is completely out of wack! I chaneged to Balance after get, and to Balance in second line...
-  getBalance() {
-    return this.http.get<Balance[]>("????????????????");
-  }
+	// this is completely out of wack! I chaneged to Balance after get, and to Balance in second line...
+	getBalance() {
+		// return this.http.get<Balance[]>("????????????????");
+	}
 
 }
