+++ angular/SIG/src/app/app.component.html
@@ -1,3 +1,1 @@
-<h2><a routerLinkActive="active" routerLink="accounts-summary">Account Summary</a></h2>
-
 <router-outlet></router-outlet>
