+++ angular/SIG/src/app/app.module.ts
@@ -1,28 +1,38 @@
-import { appRoutes } from './routes';
-import { HttpClientModule } from '@angular/common/http'; import { BrowserModule } from '@angular/platform-browser';
+import { HttpClientModule } from '@angular/common/http';
+import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
+import { AppRoutingModule } from './app-routing.module';
+import { FormsModule } from '@angular/forms';
+
+import { BalanceService } from './Services/balance.service';
+
 import { AppComponent } from './app.component';
-import { LoginComponent } from './login/login.component';
-import { AccountdetailComponent } from './accountdetail/accountdetail.component';
-import { RegisterComponent } from './register/register.component';
+import { LoginComponent } from './Components/login/login.component';
+import { AccountsComponent } from './Components/accounts/accounts.component';
+import { RegisterComponent } from './Components/register/register.component';
+import { OrdersComponent } from './Components/orders/orders.component';
+import { HomeComponent } from './Components/home/home.component';
+import { AccountdetailComponent } from './Components/accountdetail/accountdetail.component';
 
 @NgModule({
 	declarations: [
 		AppComponent,
 		LoginComponent,
-		AccountdetailComponent,
 		RegisterComponent,
 		OrdersComponent,
 		AccountsComponent,
 		HomeComponent,
+		AccountdetailComponent,
 	],
 	imports: [
 		BrowserModule,
 		AppRoutingModule,
 		ReactiveFormsModule,
 		HttpClientModule,
-		RouterModule.forRoot(appRoutes)],
+		FormsModule,
+		ReactiveFormsModule,
+	],
 	providers: [
 		BalanceService
 	],
